BACKEND/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.6
sqlalchemy==2.0.32
pydantic==2.8.2
pydantic-core==2.20.1
python-multipart==0.0.9
pillow==10.4.0
transformers==4.43.3
torch>=2.2.0
opencv-python==4.10.0.84
numpy==1.26.4
easyocr==1.7.2
torchvision>=0.16.0
